Release v1.13.0

+++ Projects/STM32H743I-EVAL/Applications/USB_Host/FWupgrade_Standalone/Src/command.c
@@ -192,7 +192,7 @@ void COMMAND_Jump(void)
 static void COMMAND_ProgramFlashMemory(void)
 {
   uint8_t readflag = TRUE;
-  uint16_t bytesread = 0;
+  uint32_t bytesread = 0;
   
   /* RAM Address Initialization */
   RamAddress = (uint32_t) &RAM_Buf;
